feat: enhance Storybook setup and improve wallet connection handling
- Add environment variable setup for Storybook to prevent errors
- Implement WagmiAdapter for improved wallet connection management
- Refactor wallet connection modal and auto-connect components
- Update transaction queue logging for better debugging
- Remove deprecated Web3 mock decorators from stories
